switch from ecdsa key to rsa
Tried to set up MediaWiki, but apparently the OIDC extension for it
didn't support ecdsa/ES256. RSA/RS256 is also what both KTH and Google
use so it seems like a reasonable choice. The only problem is that the
env variable is now huge :)
That could probably be solved by instead storing a seed to an rng used
to generate a key on startup, but that makes startup take a little bit
longer (probably around a second) and I wouldn't be completely sure
about the securit around doing that.